+++ src/MenuBar/MenuBar.test.js
@@ -75,9 +75,7 @@ const domElements = {
 const onInit = jest.fn();
 const onStateChange = jest.fn();
 const onDestroy = jest.fn();
-const onPopupStateChange = jest.fn();
 const onPopupInit = jest.fn();
-const onPopupDestroy = jest.fn();
 const { list } = domElements;
 
 const menuBar = new MenuBar({
@@ -85,9 +83,7 @@ const menuBar = new MenuBar({
   onInit,
   onStateChange,
   onDestroy,
-  onPopupStateChange,
   onPopupInit,
-  onPopupDestroy,
 });
 
 describe('Menu collects DOM elements and adds attributes', () => {
@@ -174,7 +170,6 @@ describe('Menu correctly responds to events', () => {
       domElements.listFirstItem.focus();
       domElements.listFirstItem.dispatchEvent(keydownDown);
       expect(document.activeElement).toEqual(domElements.sublistOneFirstItem);
-      expect(onPopupStateChange).toHaveBeenCalled();
     });
 
   it('Should move focus to the first popup child with spacebar from Menu bar',
@@ -182,7 +177,7 @@ describe('Menu correctly responds to events', () => {
       domElements.listFirstItem.focus();
       domElements.listFirstItem.dispatchEvent(keydownSpace);
       expect(document.activeElement).toEqual(domElements.listFirstItem.popup.firstChild);
-      expect(domElements.listFirstItem.popup.getState().expanded).toBeTruthy();
+      expect(menuBar.getState().expanded).toBeTruthy();
     });
 
   it('Should move focus to the first popup child with return key from Menu bar',
@@ -190,7 +185,7 @@ describe('Menu correctly responds to events', () => {
       domElements.listFirstItem.focus();
       domElements.listFirstItem.dispatchEvent(keydownReturn);
       expect(document.activeElement).toEqual(domElements.listFirstItem.popup.firstChild);
-      expect(domElements.listFirstItem.popup.getState().expanded).toBeTruthy();
+      expect(menuBar.getState().expanded).toBeTruthy();
     });
 
   it('Should close the submenu on right arrow key on a menu item with no submenu', () => {
@@ -200,7 +195,7 @@ describe('Menu correctly responds to events', () => {
     domElements.sublistTwoThirdItem.focus();
     domElements.sublistTwoThirdItem.dispatchEvent(keydownRight);
     expect(document.activeElement).toEqual(domElements.listFourthItem);
-    expect(domElements.listThirdItem.popup.getState().expanded).toBeFalsy();
+    expect(menuBar.getState().expanded).toBeFalsy();
   });
 
   it('Should close the submenu on left arrow key on a menu item with no parent menu', () => {
@@ -210,7 +205,7 @@ describe('Menu correctly responds to events', () => {
     domElements.sublistTwoThirdItem.focus();
     domElements.sublistTwoThirdItem.dispatchEvent(keydownLeft);
     expect(document.activeElement).toEqual(domElements.listSecondItem);
-    expect(domElements.listThirdItem.popup.getState().expanded).toBeFalsy();
+    expect(menuBar.getState().expanded).toBeFalsy();
   });
 
   it('Should click the submenu item on spacebar or return key', () => {