+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - BREAKING: Deprecates the Menu and MenuBar `menu` config property in favor of `list` (#20)
 - BREAKING: Deprecates the Tablist `tablist` config property in favor of `tabs` (#20)
 - BREAKING: Updates the way the `componentName` and self references are managed (#21)
+- BREAKING: Deprecates MenuBar `onPopupStateChange` and `onPopupDestroy` callbacks (#22)
+- Improves tracking of internal Popup state (#22)
 
 **Added**